Rename file to `page_info.dart`
