+++ lib/store/contact.dart
@@ -23,8 +23,8 @@ import 'package:collection/collection.dart';
 import 'package:get/get.dart';
 import 'package:hive/hive.dart';
 
-import '/api/backend/extension/common.dart';
 import '/api/backend/extension/contact.dart';
+import '/api/backend/extension/page_info.dart';
 import '/api/backend/extension/user.dart';
 import '/api/backend/schema.dart';
 import '/domain/model/chat.dart';
